[improvement][headless]Move QueryCache from server to core and move yaml-related classes from core to server
