+++ headless/core/src/main/java/com/tencent/supersonic/headless/core/cache/CacheManager.java
@@ -1,4 +1,4 @@
-package com.tencent.supersonic.headless.server.cache;
+package com.tencent.supersonic.headless.core.cache;
 
 
 public interface CacheManager {
