+++ headless/server/src/main/java/com/tencent/supersonic/headless/server/manager/DimensionYamlManager.java
@@ -1,9 +1,9 @@
-package com.tencent.supersonic.headless.core.manager;
+package com.tencent.supersonic.headless.server.manager;
 
 
 import com.tencent.supersonic.headless.api.enums.IdentifyType;
 import com.tencent.supersonic.headless.api.response.DimensionResp;
-import com.tencent.supersonic.headless.core.pojo.yaml.DimensionYamlTpl;
+import com.tencent.supersonic.headless.server.pojo.yaml.DimensionYamlTpl;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.BeanUtils;
 import org.springframework.stereotype.Service;
