+++ headless/server/src/main/java/com/tencent/supersonic/headless/server/service/impl/ModelServiceImpl.java
@@ -31,12 +31,12 @@ import com.tencent.supersonic.headless.api.response.MetricSchemaResp;
 import com.tencent.supersonic.headless.api.response.ModelResp;
 import com.tencent.supersonic.headless.api.response.ModelSchemaResp;
 import com.tencent.supersonic.headless.api.response.UnAvailableItemResp;
-import com.tencent.supersonic.headless.core.manager.DimensionYamlManager;
-import com.tencent.supersonic.headless.core.manager.MetricYamlManager;
-import com.tencent.supersonic.headless.core.manager.ModelYamlManager;
-import com.tencent.supersonic.headless.core.pojo.yaml.DataModelYamlTpl;
-import com.tencent.supersonic.headless.core.pojo.yaml.DimensionYamlTpl;
-import com.tencent.supersonic.headless.core.pojo.yaml.MetricYamlTpl;
+import com.tencent.supersonic.headless.server.manager.DimensionYamlManager;
+import com.tencent.supersonic.headless.server.manager.MetricYamlManager;
+import com.tencent.supersonic.headless.server.manager.ModelYamlManager;
+import com.tencent.supersonic.headless.server.pojo.yaml.DataModelYamlTpl;
+import com.tencent.supersonic.headless.server.pojo.yaml.DimensionYamlTpl;
+import com.tencent.supersonic.headless.server.pojo.yaml.MetricYamlTpl;
 import com.tencent.supersonic.headless.server.persistence.dataobject.DateInfoDO;
 import com.tencent.supersonic.headless.server.persistence.dataobject.ModelDO;
 import com.tencent.supersonic.headless.server.persistence.repository.DateInfoRepository;
